# Optimizing the function is_double_dot_path_segment