This makes is_double_dot_path_segment quite fast. Basically, we check that the length is 2, 4 or 6 and that the string begins by '.' or '%'. With only these two conditions, we can basically avoid the rest of the processing in almost all cases.

```
This PR:
BasicBench_AdaURL       7727 ns         7727 ns        90425 time/byte=9.56277ns
BasicBench_AdaURL       7754 ns         7753 ns        90129 time/byte=9.59574ns
BasicBench_AdaURL       7844 ns         7843 ns        90574 time/byte=9.70674ns
BasicBench_AdaURL       7733 ns         7733 ns        90590 time/byte=9.57049ns
Main branch:
BasicBench_AdaURL       8178 ns         8178 ns        85829 time/byte=10.1209ns
BasicBench_AdaURL       8006 ns         8006 ns        87165 time/byte=9.90789ns
BasicBench_AdaURL       8017 ns         8016 ns        87042 time/byte=9.9212ns
BasicBench_AdaURL       7974 ns         7973 ns        87570 time/byte=9.86818ns
```